Move actions to listActions file
